+++ sdk/schemaregistry/azure-schemaregistry-avroencoder/azure/schemaregistry/encoder/avroencoder/__init__.py
@@ -28,11 +28,10 @@ from ._version import VERSION
 __version__ = VERSION
 
 from ._schema_registry_avro_encoder import AvroEncoder
-from ._message_protocol import MessageType, MessageCallbackType, MessageMetadataDict
+from ._message_protocol import MessageType, MessageMetadataDict
 
 __all__ = [
     "AvroEncoder",
     "MessageType",
-    "MessageCallbackType",
     "MessageMetadataDict"
 ]