+++ sdk/schemaregistry/azure-schemaregistry-avroencoder/azure/schemaregistry/encoder/avroencoder/aio/_schema_registry_avro_encoder_async.py
@@ -24,7 +24,7 @@
 #
 # --------------------------------------------------------------------------
 from io import BytesIO
-from typing import Any, Dict, Mapping, Union, Optional, Type
+from typing import Any, Callable, Dict, Mapping, Union, Optional, Callable
 from ._async_lru import alru_cache
 from .._constants import (
     SCHEMA_ID_START_INDEX,
@@ -33,7 +33,7 @@ from .._constants import (
     AVRO_MIME_TYPE,
     RECORD_FORMAT_IDENTIFIER_LENGTH,
 )
-from .._message_protocol import MessageType, MessageCallbackType, MessageMetadataDict
+from .._message_protocol import MessageType, MessageMetadataDict
 from .._apache_avro_encoder import ApacheAvroObjectEncoder as AvroObjectEncoder
 from ..exceptions import (
     SchemaParseError,
@@ -125,7 +125,7 @@ class AvroEncoder(object):
         data: Mapping[str, Any],
         *,
         schema: str,
-        message_type: Optional[Union[MessageCallbackType, Type[MessageType]]] = None,
+        message_type: Optional[Callable] = None,
         **kwargs: Any,
     ) -> Union[MessageType, MessageMetadataDict]:
 
@@ -145,8 +145,12 @@ class AvroEncoder(object):
         :type data: Mapping[str, Any]
         :keyword schema: Required. The schema used to encode the data.
         :paramtype schema: str
-        :keyword message_type: The callback function or message class to construct the message.
-        :paramtype message_type: MessageCallbackType or MessageType or None
+        :keyword message_type: The callback function or message class to construct the message. If message class,
+         it must be a subtype of the azure.schemaregistry.encoder.avroencoder.MessageType protocol.
+         If callback function, it must have the following method signature:
+         `(data: bytes, content_type: str, **kwargs) -> MessageType`, where `data` and `content_type`
+         are positional parameters.
+        :paramtype message_type: Callable or None
         :rtype: MessageType or MessageMetadataDict
         :raises ~azure.schemaregistry.encoder.avroencoder.exceptions.SchemaParseError:
             Indicates an issue with parsing schema.
@@ -185,13 +189,17 @@ class AvroEncoder(object):
         stream.close()
         if message_type:
             try:
-                return message_type(data=payload, content_type=content_type, **kwargs)
-            except TypeError as e:
-                SchemaEncodeError(
-                    f"""The data model {str(message_type)} does not support MessageCallbackType protocol.
-                        If using an Azure SDK model class, please check the README.md for the full list
-                        of supported Azure SDK models and their corresponding versions."""
-                ).raise_with_traceback()
+                return message_type.from_message_data(payload, content_type, **kwargs)
+            except AttributeError:
+                try:
+                    return message_type(payload, content_type, **kwargs)
+                except TypeError as e:
+                    SchemaEncodeError(
+                        f"""The data model {str(message_type)} is not a Callable that takes `data`
+                            and `content_type` or a subtype of the MessageType protocol.
+                            If using an Azure SDK model class, please check the README.md for the full list
+                            of supported Azure SDK models and their corresponding versions."""
+                    ).raise_with_traceback()
 
         return {"data": payload, "content_type": content_type}
 
@@ -245,7 +253,7 @@ class AvroEncoder(object):
                 content_type = message.__content_type__()
             except AttributeError:
                 SchemaDecodeError(
-                    f"""The data model {str(message)} does not support MessageType protocol.
+                    f"""The data model {str(message)} is not a subtype of the MessageType protocol.
                         If using an Azure SDK model class, please check the README.md for the full list
                         of supported Azure SDK models and their corresponding versions."""
                 ).raise_with_traceback()