+++ sdk/schemaregistry/azure-schemaregistry-avroencoder/samples/sync_samples/encode_and_decode_with_metadata.py
@@ -60,9 +60,9 @@ def encode_metadata_dict(encoder):
     encoded_metadata_ben = encoder.encode(dict_data_ben, schema=SCHEMA_STRING)
 
     print("Encoded metadata is: ", encoded_metadata_ben)
-    return EventData(
-        data=encoded_metadata_ben["data"],
-        content_type=encoded_metadata_ben["content_type"],
+    return EventData.from_message_data(
+        encoded_metadata_ben["data"],
+        encoded_metadata_ben["content_type"],
     )
 
 def decode_with_data_and_content_type(encoder, event_data):